(VANAGON-56) Introducing Mirror URLs
The rewrite engine in Vanagon is not popular. It's made hitting
internal mirrors easier for Puppet but it's made building Puppet Agent
harder for many of our customers. I've been planning on adding Mirror
URL support for a while now... and that day is today.

Mirror URLs will deprecate rewrite rules, and provide flexibility for
multiple mirrors (if so desired). There's a new DSL method, and some
backend code, and deprecated-but-preserved backwards compatibility for
any projects that aren't likely to be able to move forward quickly.