(maint) Add ability to install service as a link
Some packaging installs service files as links to executables. This
commit adds the `link_target` keyword argument to install_service. If
`link_target` is specified the specified `service_file` will be
installed at the location in `link_target` and the target service file
will be a symlink to the `link_target`.