(VANAGON-55) Fix AIX RPMs requirements
AIX doesn't track the entire filesystem in the RPM database, so
there's a number of core system files that are not exposed to RPM and
should not be used as a constraint for dependencies. `/bin/touch` and
`/bin/mkdir` are two of them. We should stop requiring those files as
dependencies for AIX RPMs.