(maint) Fix for workdir with symlinks
Fun fact: When building on OSX your tmpdir is going to end up in /var
which is actually a symlink to /private/var. If you don't chdir / set
your workdir to the realpath of your tmpdir git will fail with

fatal: Could not chdir to
'../../../../../../../../../../../private/var/folders...'

if you are using the `--git-dir` option. The ruby-git library we use
appends `--git-dir` to each git command. So, when initializing the git
objects we need to make sure we're using the `File.realpath` so we don't
trigger this weird issue when trying to update submodules.

Big thanks to @mckern for helping me track this issue down.