(maint) Do not create any backup files on patch
Prior to this commit, if a patch wasn't perfectly cleanly applied, the
patch command would create a .orig file. This often would happen if
the patch was offset by some amount. This commit updates the patch
process to prevent that backup file from being created. We don't need
that, given that we're working on a throwaway system. If the patch can't
apply, it'll throw an error. If it does apply, even if there's an offset
issue, we don't need to save that backup file.