(RE-8839) Stage packages in plat-tag directory
This commit updates how built packages are staged. Rather than being
opinionated on repo or path structure for specific platforms, this
commit changes the default behavior of vanagon to simply stage the
packages under the platform tag (platform-version-architecture). This
allows us to concentrate the automation that is opinionated about path
structure and repo formats into puppetlabs/packaging, which comes into
play when shipping out the packages.