(maint) Simplify compiled archive generation
The way we were populating the compiled archives in the past caused
failures on Solaris. The version of `gtar` on Solaris 10 and 11 claims
to support the `-C` flag to change directories prior to adding files,
however this ends up giving errors about the directories not being
found. Since each line of the makefile executes in a separate shell
anyways, let's just `cd` prior to generating the tarball to simplify the
commands.