(maint) Do not place source URIs into the mirror list
When populating the mirror list, vanagon attempts to include rewritten
URIs as a fallback for when the user has not specified any mirrors
explicitly.

Unfortunately, the URI rewriter returns the original URI if there is
no matching rule. This means that the original source URI ends up in
the mirror list, and can potentially be accidentally used even when
the user has specified a dedicated mirror to target.