Fix Style/MixinUsage warning
We were including vanagon/utilities in a weird place halfway through
the call stack, so anything that was loaded after platform/rpm.rb
inherited them, but anything before may not. The inclusion of
vanagon/utilities has been made explicit and occurs at what is
probably the right depth in the stack.