(VANAGON-116) Retry source fetches individually
Previously, Vanagon would retry fetching all sources as a single
block. This could confuse debugging of failed fetches, since it's
possible for a different source to fail in the retry than in the
original run. It also (paradoxically) increased the chances for an
overall job failure, since successful downloads were given another
chance to break.

This change causes Vanagon to retry fetching each component
individually, giving each an opportunity to succeed individually, and
making debugging cleaner when fetches fail.

Since on most platforms build dependencies are installed in a single
package manager transaction, that is still treated monolithically in
the driver.