(PA-1245) Add support for no_packaging and fetch_artifact
This commit adds a project object function to specify that the project is
going to fetch additional files after the build.

Additionally, support has been added for the 'no_packaging' setting, that
identifies that the project should not perform packaging operations and just
stop after the installation steps have completed.