(VANAGON-120) Run postinstall scripts in the postinstall phase
Previously, Vanagon ran postinstall scripts for upgrades in the
postuninstall phase. This meant that any postinstall scripts would
always be sourced from the /previous/ version of a package,
which makes it rather difficult to fix bugs in the script or to
change the contract between a preinstall and postinstall phase.

This runs those scripts in the expected phase of the RPM
transaction.