(VANAGON-128) Add build requirement overrides
This commit adds a new feature to the platform object:
platform.override_build_requirement specifies that the requirement given as a
paramter should be overriden and instead use the override when a component
specifies the original requirement.

For example, if a component specifies:

component.build_requires 'gcc'

and the platform specifies

platform.override_build_requirement 'gcc', 'pl-gcc'

the component will prefer the override and specify 'pl-gcc' as the actual
requirement.

This functionality can be ignored in the component by setting
ignore_overrides: true like so:

component.build_requires 'gcc', ignore_overrides: true