(PA-1991) Allow inheriting settings from a local project directory
Adds an `inherit_local_settings` function to the DSL that allows vanagon
projects to inherit settings from a vanagon project in a separate local
directory. Accepts a vanagon project name and a path to the project
directory. Does not offer the ability to specify a git revision - the
files are used as they are found.