(VANAGON-131) Optionally load project settings from yaml
Adds an `inherit_yaml_settings` method to the project DSL that allows
projects to specify a URL to a yaml file containing settings as
generated by `publish_yaml_settings`. If the settings URL uses http, a
URL for a corresponding sha1sum file must be supplied as well. For file
URLs, the settings URL by itself is sufficient.