(maint) Add the option to build a Vanagon project with Docker
This commit introduces the build_with_docker option to Vanagon's build
command that, when specified, will configure Vanagon to build the
project with Docker via. a generated Dockerfile. It does _not_, however,
generate any packaging artifacts. Support for that will be added in a
later commit.

Building the project with Docker has many benefits. Perhaps the most
important one is that we get Docker's caching for free. This means that
if someone modifies a .rb file of a downstream component and then tries
to rebuild the project, Vanagon (Docker) will resume the build at the first
changed instruction in the Dockerfile instead of starting the build from
the beginning, which is what it currently does for the default
Makefile-based builds.

In order to use this feature, platform.rb files must define a
base_docker_image attribute.

Existing Vanagon projects that would like to take advantage of this
feature must ensure that none of the component commands use
Make-specific syntax. Common Make-specific syntax includes:
* Variable references of the form $(var). Use ${var} instead.
* Shell expressions of the form $(shell expr <expr>)
* Shell commands of the form $(shell <command>)
* $$

Note that if eliminating the Make-specific syntax is not possible, then
some additional work may need to be done to ensure that Vanagon projects
can be configured to be built with Docker or Make. This work can be done
separately from the changes here.