(maint) Address PR review comments
This commit contains the following changes:

* Change the default remote working directory from /root => /build

* Change build_with_docker => build-with-docker

* Use puts instead of warn

* Change to_dockerfile => dockerfile to keep things consistent with the
existing rules method

* Invoke the individual <step_name>_step methods as <step_name>_step()
in Component::Dockerfile (i.e. invoke them with the parenthesis). This
makes it clear that these are method calls and _not_ a list of functions.

* Validate that docker's present on localhost at an earlier step when
building the project with Docker

* Allow one to specify the image tag in Vanagon::Docker.build

* Tag the built Docker image as "<project_name>-<git_describe>:latest"

* Copy each component's sources at that component's build step so that
we can take advantage of caching when e.g. a downstream component's source
changes or when the user's building the project with the --only_build option.

* Create each component's directories at that component's build step for
the same reason