(PE-24814) Add handling for ghost files in rpm specs
Making use of the alternates service in a Redhat platform requires
tracking target files as 'ghost' files in the rpm spec.

Per https://fedoraproject.org/wiki/Packaging:Alternatives#How_to_use_alternatives
we touch the generic links so that rpm will be aware of them when
building the package, but mark them now as %ghost in the %files list
so that the dummy files we touched are not actually copied in when the
package is installed.

This is intended to allow commands like `rpm -qf
/opt/puppetlabs/server/bin/psql` to correctly report that the symlink
(created in the post trigger by update-alternatives) is managed by
the pe-postgresql10 package, for example...

This patch adds some getters and setters to allow adding these ghost
file settings through the Component::DSL and then having the Project add
them to the final rpm spec.