(VANAGON-146) Improve parsing of project vendor field
This commit replaces several instances of regex matching on the vendor field
with methods `vendor_name_only` and `vendor_field_only` to parse out either just
the name or just the email address, respectively, from a vendor value of the
form 'Name of Vendor <email@vendor.com>'.
This commit also adds a `vendor=` setter to the `Vanagon::Project` class in
order to validate that the vendor value is of the appropriate form. Previously,
if the email portion was not included, erb template parsing would fail with
"undefined method `[]' for nil:NilClass". Now we provide a helpful error message
if the vendor value is not of the correct form.