(VANAGON-101) Sanitize docker container names
Docker requires container names to match the following regular expression:

    [a-zA-Z0-9][a-zA-Z0-9_.-]

Vanagon will violate this when it attempts to launch a container from an
image like `amd64/debian:9-slim` as the container is named by appending
`-builder` to the image name. This commit updates the Docker engine to
sanitize the name given to Docker containers by replacing `/` and `:`
characters with underscores.