(PA-2670) Add support for `get_version_forced` in the vanagon component
This will allow us to force an additional clone of git sources to
determine the version while the component is still being processed. This
is needed to allow for including the component version in the gemspec
file for facter, puppet, and hiera so that ruby correctly identifies
those as installed gems.