(VANAGON-153) Update version in SMF manifest
When upgrading the puppet-agent package, Solaris 11 doesn't
automatically restart a service if the SMF manifest contains no changes.

This commit runs a sed on the target service(s) to change the current
version (which is always 1) to the puppet-runtime version (since version
is expected to be an integer value, per SMF documentation).

This commit also changes the Solaris sed to /usr/gnu/bin/sed which is
the GNU implementation.