(VANAGON-157) Make curl be more clear about fetch errors.
This commit adds the '--silent' '--show-error' and '--fail' flags to
calls made by vanagon to force bad fetches to return non-zero exit
status in the shell.

As a side-effect, some of the curl output statistics is silenced. I
think this silencing is desirable.

The previous behavior looked like this upon success:

$ curl -o STUFF https://artifactory.delivery.puppetlabs.net/artifactory/scratchpad__local/exg-test/foo.txt
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
100    29  100    29    0     0    159      0 --:--:-- --:--:-- --:--:--   160
$ echo $?
0

The previous behavior looked like this upon  failure:

$ curl -o STUFF https://artifactory.delivery.puppetlabs.net/artifactory/scratchpad__local/exg-test/foo.txt.BAD
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
100    80    0    80    0     0    419      0 --:--:-- --:--:-- --:--:--   421
$ echo $?
0

The new behavior looks like this upon success:

$ curl --silent --show-error --fail -o STUFF https://artifactory.delivery.puppetlabs.net/artifactory/scratchpad__local/exg-test/foo.txt
$ echo $?
0

The new behavior looks like this upon failure:

$ curl --silent --show-error --fail -o STUFF https://artifactory.delivery.puppetlabs.net/artifactory/scratchpad__local/exg-test/foo.txt.BAD
curl: (22) The requested URL returned error: 404 Not Found
$ echo $?
22