(RE-13396) Fix debug_package handling in rpmbuilds
The version we had:
    %global debug_package %{nil}
was working fine until last night but due to the vagueries of RPM macro handling, it
stopped.

The net suggests that the correct invocation is:
    %define debug_package %{nil}

This is what I've put in and it works in my tests. The difference of %define and %global
are both scoping and lazy evaluation of the macro body.

I chose to go with the "standard" version but this also works:
    %global debug_package %%{nil}

The above makes the 'global' context but also forces lazy-evalution of %{nil}

I am completely perplexed why either the scope or the evaluation of %{nil} should be an issue
here. My speculative guess is that there's now a bug in rpmbuild.

Nevertheless, the "standard" invocation should work across all platforms.