(maint) Output valid JSON build metadata
As part of the build process we run `echo -e "#{metadata}" >
output/foo.json` where `metadata` is a JSON string. This doesn't
consistently preserve double quotes, causing the resulting file to be
invalid JSON. This PR escapes the double quotes in the metadata before
echoing to ensure the resulting json file can be consumed.