(VANAGON-163) Short-circuit engine selection logic
Vanagon has a bunch of logic for picking an engine based on the
configuration of the platform and other bits of info. This is
all well and good, but the logic was applied even when a user
had explicitly passed the `--engine` option to the CLI.

This commit updates the Vanagon::Driver to skip the selection
logic when an engine has been explicitly chosen.