(VANAGON-163) Provide error context for engine load failure
Prior to this commit, the error handling in `Vanagon::Driver#load_engine_object`
would fail with a generic "No such engine" message. This was misleading as
the engine often _does exist_ and loading failed due to a specific issue.
For example, `Vanagon::Engine::Docker` cannot load if the `docker` binary
is not available on the `PATH`. This commit updates the error trap to
preserve context by using `Vanagon::Error.wrap` and also consolidates
error handling in `Vanagon::Driver#initialize` that trapped instances
of `LoadError` by catching the more generic `ScriptError` in addition
to `StandardError`.