(DIO-1066) Update vanagon to go through ABS for VMs Before this change, the ABS engine was partially implemented. It now includes the select_target and teardown methods needed for vanagon build lifecycle.
This engine is thus a level of abstraction that supports any downstream
engine (vmpooler, nspooler, AWS) and is forward compatible with
ondemand Vmpooler pools.

The ABS token is read from env var ABS_TOKEN or the first line of ~/.vanagon-token
or the vmfloaty config (using top-level token or the first abs type service found)

The vmpooler token optionaly can be provided through an env var VMPOOLER_TOKEN
or the second line of ~/.vanagon-token or via vmfloaty's existing vmpooler_fallback config

The ABS request is done and checked for up to an hour, with feedback given on the
availability of resources. Teardown is done at the end using the job_id.

Users can track vm usage via bitbar or floaty if they provide a vmpooler token.