(maint) Overwrite unchanged config files
Configuration files installed with `install_configfile` have the
`.pristine` extension appended to them, as a guard to avoid clobbering
existing configfiles.

OS X and Solaris 10 use postinstall scripts to manage configuration
files, removing the extension if a non-pristine config file does not
exist.

However, we don't need to keep the pristine files if there are no
differences between the pristine file and the target file. So, update
the logic to also run a `diff` between the two files. If there are no
differences, go ahead and override the target file.