(maint) Fix pristine configfile clobbering
On Solaris 10, postinstall scripts are executed by the `sh` interpreter,
as the package manager does not seem to respect the shebang line at the
top of the file.

This causes the `!` sign to not be recognized, and the configfiles are
clobbered regardless of the differences between the installed config
file and the pristine config file.

Sample output after upgrading puppet-agent on Solaris 10:

```
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
/var/sadm/pkg/puppet-agent/install/postinstall: !: not found
root:x:0:0:Super-User:/:/usr/bin/bash
root::0:
Installation of <puppet-agent> was successful.
```

To avoid these types of errors, do not use `!` in the script, and remove
the pristine files instead of tampering with the non-pristine ones.

This also takes care of the case when `diff` is not present on the
system, causing the check to noop instead of clobbering.