(RE-13837) Hack around Git.ls_remote failure
Create a slightly sloppy workaround for a failure in Git.ls_remote

When Git.ls_remote calls the git command-line, under moderately common
circumstances, ssh can be involved and it can print something like:

    Warning: Permanently added 'github.com,192.30.255.113' (RSA) ...

Git.ls_remote, does not anticipate this and attempts to parse it as
actual `git ls-remote` output.

This results in a `NoMethodError: undefined method `split' for
nil:NilClass` because the
[#split][https://github.com/ruby-git/ruby-git/blob/bd026d39b17e192e8a710744153ed28139cc64a7/lib/git/lib.rb#L502]
fails on the unexpected message.

I think that leaves us with 3 possible solutions.

1. Not use ruby-git for this operation (which was my choice).
1. monkey-patch the Git.ls_remote method (which was my 2nd choice)
1. catch the 'NoMethodError' exception and try to sort it out (yuck!)

It would be good citizenry to submit a PR to ruby-git which I might do
to correctly fix this problem. I may endeavor to do so.