(PA-3570) enable clone into a custom dirname
puppet-agent#6.x needs to ship both facter#3.x and
facter#main branches using 2 different components(
facter#3.x -> facter, facter#main -> facter-ng).
In order to do this, the component must configure
the directory where the git repository should be cloned.

This PR makes use of `dirname` attribute that can
be set at component level and exposes this to the
Git source in order to enable the ability to clone
and use a repository with a specific name.