(RE-14154) Exit if `VANAGON_FORCE_SIGNING` is set
By default, Vanagon does not fail if extra files signing fails, it just
logs an error and continues building the package. This is unwanted
behavior in environments where we expect a hard failure when signing
cannot proceed.

To force Vanagon to fail if extra files signing fails, ensure the
VANAGON_FORCE_SIGNING variable is set before starting a build.