(maint) Add brew method for macOS
In 2020, Apple began releasing machines that used new ARM-based
processors, as opposed to the x86 processors it had used for many years
prior.

Homebrew, a macOS package manager, changes its path depending on whether
it's running on x86-64 or ARM64.

This commit adds a new `brew` method to Vanagon that defaults to the x86
path (/usr/local/bin/brew) but can be overriden to the ARM path
(/opt/homebrew/bin/brew).