(VANAGON-214) Create `test` home directory before running brew
Previously, if brew install tried to log to `HOMEBREW_LOGS`, which defaults to
`~/Library/Logs`, then it would fail because the directory didn't exist yet.

Now, explicitly create the `test` home directory before installing `brew`. Since
the OS creates the directory, it creates the appropriate skeleton
directories/files and sets the permissions so it's owned by the `test` user.