(maint) Fix git remote detection
Previously, this was attempting to determine if a github.com URL was a remote or a media type by inspecting part of the path. However, this was missing paths with 'release' in them. Also, we were never actually using any other return type but :github_remote. This changes the logic to return :github_media if the path contains known media tokens, or if the suffix ends with .tar.gz or .zip. Otherwise, we assume it's a remote.

Ideally, we'd reach out to GitHub to determine if it's a repo, but we're doing this to avoid rate limiting.