(PA-5786) Stop using posttrans for post-install steps
Before this commit, vanagon used the `%posttrans` section of an
rpm spec for #add_postinstall_action. Specifying items in the
`%posttrans` section means that rpm considers the transaction
done, and rpm continues to excecute irregardless of the result
of the `%posttrans` scriptlet. This behavior breaks puppet-agent
8 fips upgrades because certain configuration files are created
in a #add_postinstall_action, and those configuration files were
not yet ready before the agent daemon is killed and then started
as part of the upgrade.

This change eliminates the `%posttrans` section entirely in favor
of moving all #add_postinstall_action code to the `%post` section.