(PA-5786) Add ability to execute direct post installation scriptlets
The existing `#add_postinstall_action` adds actions to the `%posttrans`
section of an rpm spec file. This can mean that the action could fail,
but the installation/upgrade could still succeed. This commit adds the
ability to add actions that must succeed post installation for rpm
installations by adding the action to the `%post` section.