(PA-6886) Add Digicert to Solaris images
The DigiCertGlobalRootG2 is now used by Artifactory and the VMs we're using
for Solaris don't contain that cert by default.
We for sure update our images to include that, but until that is done we need
to make sure to use it.
Also, in our case the OpenSSL used on our Solaris VMs is ancient and doesn't support
openssl-rehash, so we have to create the sim link ourself.
Once we update our images this commit can be reverted.